requirements.txt
# Core ML Libraries
pandas>=2.0.0
numpy>=1.21.0
scikit-learn>=1.3.0
scipy>=1.9.0

# Deep Learning (optional)
tensorflow>=2.12.0
torch>=2.0.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# ML Pipeline & Orchestration
mlflow>=2.5.0
prefect>=2.10.0
hydra-core>=1.3.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Data Processing
openpyxl>=3.1.0
xlrd>=2.0.0

# Model Interpretation
shap>=0.42.0
lime>=0.2.0

# Statistical Analysis
statsmodels>=0.14.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Jupyter & Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
nbconvert>=7.7.0
